+++ packages/seacas/libraries/ioss/src/Ioss_Utils.C
@@ -1514,7 +1514,7 @@ int Ioss::Utils::log_power_2(uint64_t value)
   value |= value >> 8;
   value |= value >> 16;
   value |= value >> 32;
-  return tab64[((uint64_t)((value - (value >> 1)) * 0x07EDD5E59A4E28C2)) >> 58];
+  return tab64[(((value - (value >> 1)) * 0x07EDD5E59A4E28C2)) >> 58];
 }
 
 void Ioss::Utils::copy_database(Ioss::Region &region, Ioss::Region &output_region,
